requirements.txt
requests>=2.28.0
python-binance>=1.0.16
python-dotenv>=0.19.0
